database-ai-agents-main/requirements.txt
# Database and data processing
pyodbc==5.2.0
SQLAlchemy==2.0.36
pandas==2.2.3
tabulate==0.9.0

# OpenAI and LangChain
openai==1.58.1
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.3
langchain-openai==0.2.14

# Anthropic (Claude)
langchain-anthropic==0.2.4

# Utilities
python-dotenv==1.0.1

# Web framework
streamlit==1.41.1
